refactor: shop 정보 받는 input 컴포넌트로 빼내기
